pyproject.toml
[tool.poetry]
name = "downloadsort"
version = "0.2.0"
description = ""
authors = ["ActuallySin"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
PyYAML = "^6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
